renamed game_cfg to make tab-completion easier
