+++ config/colors.h
