+++ config/game_internals.h
