+++ game.h
@@ -1,7 +1,7 @@
 #ifndef GAME_H
 #define GAME_H
 
-#include "game_cfg/game_cfg.h"
+#include "config/game_cfg.h"
 
 extern int keyboard_x, keyboard_y;
 extern char message_banner[MESSAGE_LENGTH];
